requirements.txt
setuptools==60.5.0
tldextract==3.1.2
boto3==1.20.32
dnspython==2.1.0
azure-identity==1.7.1
azure-mgmt-trafficmanager==0.51.0
python-dotenv==0.19.2
